backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
flask-pymongo
pymongo
python-dotenv
numpy<2.0
pandas
tensorflow
celery
redis
faster-whisper
requests

gunicorn
flask-sqlalchemy
psycopg2-binary
